Chloroplast alignment

Amino acid alignment of 15 genes of 19 different chloroplast.

Examples

data(chloroplast)
chloroplast
#> 19 sequences with 5144 character and 2775 different site patterns.
#> The states are A R N D C Q E G H I L K M F P S T W Y V