phangorn: Phylogenetic Reconstruction and Analysis

Allows for estimation of phylogenetic trees and networks using Maximum Likelihood, Maximum Parsimony, distance methods and Hadamard conjugation (Schliep 2011). Offers methods for tree comparison, model selection and visualization of phylogenetic networks as described in Schliep et al. (2017).

See also

Author

Maintainer: Klaus Schliep klaus.schliep@gmail.com (ORCID)

Authors:

Emmanuel Paradis (ORCID)

Leonardo de Oliveira Martins (ORCID)

Alastair Potts

Iris Bardel-Kahr (ORCID)

Other contributors:

Tim W. White [contributor]

Cyrill Stachniss [contributor]

Michelle Kendall m.kendall@imperial.ac.uk [contributor]

Keren Halabi [contributor]

Richel Bilderbeek [contributor]

Kristin Winchell [contributor]

Liam Revell [contributor]

Mike Gilchrist [contributor]

Jeremy Beaulieu [contributor]

Brian O'Meara [contributor]

Long Qu [contributor]

Joseph Brown (ORCID) [contributor]

Santiago Claramunt (ORCID) [contributor]