Read Amino Acid Sequences in a File

This function reads amino acid sequences in a file, and returns a matrix list of DNA sequences with the names of the taxa read in the file as row names.

Usage

read.aa(file, format = "interleaved", skip = 0, nlines = 0,
  comment.char = "#", seq.names = NULL)

Arguments

file

a file name specified by either a variable of mode character, or a double-quoted string.

format

a character string specifying the format of the DNA sequences. Three choices are possible: "interleaved", "sequential", or "fasta", or any unambiguous abbreviation of these.

skip

the number of lines of the input file to skip before beginning to read data.

nlines

the number of lines to be read (by default the file is read until its end).

comment.char

a single character, the remaining of the line after this character is ignored.

seq.names

the names to give to each sequence; by default the names read in the file are used.

Value

a matrix of amino acid sequences.

References

https://en.wikipedia.org/wiki/FASTA_format

Felsenstein, J. (1993) Phylip (Phylogeny Inference Package) version 3.5c. Department of Genetics, University of Washington. https://phylipweb.github.io/phylip/

See also

Author

Klaus Schliep klaus.schliep@gmail.com