Laurasiatherian data (AWCMEE)

Laurasiatherian RNA sequence data

Source

Data have been taken from the former repository of the Allan Wilson Centre and were converted to R format by klaus.schliep@gmail.com.

Examples

data(Laurasiatherian)
str(Laurasiatherian)
#> List of 47
#>  $ Platypus  : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ Wallaroo  : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Possum    : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Bandicoot : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Opposum   : int [1:1605] 2 1 4 1 3 3 4 4 4 3 ...
#>  $ Armadillo : int [1:1605] 2 1 2 1 3 3 4 2 4 3 ...
#>  $ Elephant  : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Aardvark  : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ Tenrec    : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ Hedghog   : int [1:1605] 1 4 1 1 3 3 4 2 4 3 ...
#>  $ Gymnure   : int [1:1605] 1 1 1 1 3 3 4 2 4 3 ...
#>  $ Mole      : int [1:1605] 1 1 1 1 3 3 4 4 4 3 ...
#>  $ Shrew     : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ Rbat      : int [1:1605] 2 1 4 1 3 3 2 4 4 2 ...
#>  $ FlyingFox : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ RyFlyFox  : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ FruitBat  : int [1:1605] 4 1 4 1 3 3 4 4 4 3 ...
#>  $ LongTBat  : int [1:1605] 2 1 4 1 3 3 4 4 4 3 ...
#>  $ Horse     : int [1:1605] 2 1 4 1 3 3 2 4 4 3 ...
#>  $ Donkey    : int [1:1605] 2 1 4 1 3 3 2 4 4 3 ...
#>  $ WhiteRhino: int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ IndianRhin: int [1:1605] 2 1 4 1 3 3 4 4 4 3 ...
#>  $ Pig       : int [1:1605] 2 1 2 1 3 3 4 4 4 3 ...
#>  $ Alpaca    : int [1:1605] 2 1 2 1 3 3 4 4 4 1 ...
#>  $ Cow       : int [1:1605] 2 1 4 1 3 3 4 4 4 3 ...
#>  $ Sheep     : int [1:1605] 4 1 4 1 3 3 4 4 4 3 ...
#>  $ Hippo     : int [1:1605] 4 1 2 1 3 3 4 4 4 3 ...
#>  $ FinWhale  : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ BlueWhale : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ SpermWhale: int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Rabbit    : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ Pika      : int [1:1605] 1 1 4 1 3 3 4 4 4 3 ...
#>  $ Squirrel  : int [1:1605] 2 1 4 1 3 3 4 4 4 3 ...
#>  $ Dormouse  : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ GuineaPig : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Mouse     : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Vole      : int [1:1605] 1 1 1 1 3 3 4 4 4 3 ...
#>  $ CaneRat   : int [1:1605] 2 1 1 1 4 3 4 4 4 3 ...
#>  $ Baboon    : int [1:1605] 1 1 4 1 3 3 2 4 4 3 ...
#>  $ Human     : int [1:1605] 1 1 4 1 3 3 4 4 4 3 ...
#>  $ Loris     : int [1:1605] 2 1 2 1 3 3 4 4 4 3 ...
#>  $ Cebus     : int [1:1605] 2 1 2 1 3 3 4 4 4 3 ...
#>  $ Cat       : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ Dog       : int [1:1605] 4 1 1 1 3 3 4 4 4 3 ...
#>  $ HarbSeal  : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ FurSeal   : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  $ GraySeal  : int [1:1605] 2 1 1 1 3 3 4 4 4 3 ...
#>  - attr(*, "class")= chr "phyDat"
#>  - attr(*, "weight")= int [1:1605] 1 5 1 444 3 314 1 1 328 1 ...
#>  - attr(*, "nr")= int 1605
#>  - attr(*, "nc")= num 4
#>  - attr(*, "index")= int [1:3179] 1 2 3 4 5 6 7 8 9 6 ...
#>  - attr(*, "levels")= chr [1:4] "a" "c" "g" "t"
#>  - attr(*, "allLevels")= chr [1:18] "a" "c" "g" "t" ...
#>  - attr(*, "type")= chr "DNA"
#>  - attr(*, "contrast")= num [1:18, 1:4] 1 0 0 0 0 1 1 1 0 0 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : NULL
#>   .. ..$ : chr [1:4] "a" "c" "g" "t"