Auxiliary for Controlling Fitting

Auxiliary functions for optim.pml fitting. Use it to construct a control or ratchet.par argument.

Usage

pml.control(epsilon = 1e-08, maxit = 10, trace = 1, tau = 1e-08)

ratchet.control(iter = 20L, maxit = 200L, minit = 100L, prop = 1/2)

Arguments

epsilon

Stop criterion for optimization (see details).

maxit

Maximum number of iterations (see details).

trace

Show output during optimization (see details).

tau

minimal edge length.

iter

Number of iterations to stop if there is no change.

minit

Minimum number of iterations.

prop

Only used if rearrangement=stochstic. How many NNI moves should be added to the tree in proportion of the number of taxa.´

control

A list of parameters for controlling the fitting process.

Value

A list with components named as the arguments.

Details

pml.control controls the fitting process. epsilon and maxit are only defined for the most outer loop, this affects pmlCluster, pmlPart and pmlMix. epsilon is defined as (logLik(k)-logLik(k+1))/logLik(k+1), this seems to be a good heuristics which works reasonably for small and large trees or alignments. If trace is set to zero than no out put is shown, if functions are called internally than the trace is decreased by one, so a higher of trace produces more feedback.

See also

Author

Klaus Schliep klaus.schliep@gmail.com

Examples

pml.control()
#> $epsilon
#> [1] 1e-08
#> 
#> $maxit
#> [1] 10
#> 
#> $trace
#> [1] 1
#> 
#> $tau
#> [1] 1e-08
#> 
pml.control(maxit=25)
#> $epsilon
#> [1] 1e-08
#> 
#> $maxit
#> [1] 25
#> 
#> $trace
#> [1] 1
#> 
#> $tau
#> [1] 1e-08
#>